Added the ability to define the topic type.
By default, the script uses the base DITA topic, as it is most
permissivea and generally validates without having to adjust the
converted content in any way. If the {_mod-docs-content-type} attribute
is defined in the file, the script then adds its value to the <topic>
tag as an outputclass attribute.

With this update, users can define the value of the {dita-topic-type}
attribute on the command line to change the topic type of the generated
output to concept, reference, or task.

Example: asciidoctor -a dita-topic-type=concept ...