Added support for floating and block titles.
Because DITA does not allow the <title> element to be assigned to many
blocks, I chose to use a paragraph with bold text and a proper output
class to simulate this behavior in the interest of not losing data
between the conversions. To turn this functionality off, users can
set the {dita-topic-titles} attribute value on the command line to
'strict'.

Example: asciidoctor -a dita-topic-titles=strict ...