Ensured that generated topics have assigned id.
When id is not provided in the source AsciiDoc document, the conversion
plug-in falls back to using the file name to ensure the topic has an id
assigned as DITA requires. This however did not work in situations when
the AsciiDoc content was read from STDIN and thus no file name was
available. If id is not provided and cannot be derived from the file
name, the plug-in now generates a UUID.