Stopped manipulating nodes as raw strings.
This looked like a bad idea even when I initially implemented this,
but the side effect is even worse than I anticipated. The motivation for
this change was to remove the workaround lines for forcing a role in
elements that AsciiDoc does not support. Since they are not visible in
output, the better option is to leave them an and possibly clean them up
at the post-processing stage.